Rename measure/ folder to measure2
