+++ src/models/measure2/base.cpp
@@ -1,7 +1,7 @@
 
 
 
-#include <fullscore/models/measure/base.h>
+#include <fullscore/models/measure2/base.h>
